CWW-1721 Introduce components for cards
