+++ src/components/memory/ImageCard.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const ImageCard = (props) => {
+    return (
+        <img src={props.info.image} alt="" ></img>
+    )
+}
+
+export default ImageCard
